chore: trim e2e and editor code without changing behavior
Remove redundant test assertions and helpers, simplify section device-type
selectors and the otter-store action types, and shorten the OpenAI e2e stub.
